libfusion/shm: use s64 instad of int to calculate number of free bytes to prevent wraparound when having more than 2GB free
In ubuntu 11.10 the following tmpfs filesystems are mounted by default

none                  5.0M     0  5,0M   0% /run/lock
none                  3.9G  400K  3,9G   1% /run/shm

This led to a wraparound in the 32-bit signed value for the 3.9G tmpfs
and /run/lock with its 5M was favoured.

+++ lib/fusion/shm/shm.c
@@ -77,7 +77,7 @@ fusion_find_tmpfs( char *name, int len )
              (!strcmp( mount_fs, "tmpfs" ) || !strcmp( mount_fs, "shmfs" ) || !strcmp( mount_fs, "ramfs" )))
           {
                struct statfs stat;
-               int           bytes;
+               s64           bytes;
 
                if (statfs( mount_point, &stat )) {
                     D_PERROR( "Fusion/SHM: statfs on '%s' failed!\n", mount_point );
